backend/requirements.txt
Django==5.2.12
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-environ==0.13.0
celery==5.6.2
redis==7.3.0
psycopg[binary]==3.3.3
cryptography==45.0.2
